requirements.txt
PyYAML==6.0
jupyterlab==3.2.9
numpy==1.19.5
tensorflow==2.6.2
tqdm==4.64.0
scikit-learn==0.24.2
scipy==1.5.4
pandas==1.1.5
matplotlib==3.3.4
